Refactor transcription serve into dedicated module
