+++ sglang_omni/serve/openai_errors.py
@@ -1,0 +1,19 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Shared OpenAI-compatible API error classification helpers."""
+
+from __future__ import annotations
+
+_BAD_REQUEST_MARKERS = (
+    "longer than the model's context length",
+    "Requested token count exceeds the model's maximum context length",
+    "accepts audio up to",
+    "max_new_tokens must be",
+    "exceeds the maximum allowed length",
+    "sequence exceeds max_length",
+    "multimodal_train_inputs",
+)
+
+
+def is_bad_request_error(exc: Exception) -> bool:
+    message = str(exc)
+    return any(marker in message for marker in _BAD_REQUEST_MARKERS)
